refactor(core): widen hook addresses to 64 bits
A hook address was a uint, which held while every target was 32-bit. A macOS
image sits above a 4 GB __PAGEZERO, so its addresses start at 0x100000000, and
the parser's narrowing cast turned one such address into 0x004EB6C2: a
plausible-looking address somewhere else in the same image. The validator's
single window rejected the range outright anyway.

Addresses are now ulong from the TOML through to the image seam, and the
validator carries a second window for a 64-bit Mach-O.

- src/KPatchCore/Parsers/HooksParser.cs: TryGetUInt becomes TryGetULong and
  refuses a value that is not a non-negative integer instead of casting it.
- src/KPatchCore/Parsers/PeExecutableImage.cs: these PE builds are 32-bit, so an
  address that does not fit one is refused rather than narrowed.

Every hook file in the tree still parses: 223 hooks across 74 files, no parse
failures. The eleven that fail validation are macOS files still written as file
offsets rather than absolute addresses, and they failed the old window too,
which only ever narrowed.

+++ src/KPatchCore/Parsers/ElfExecutableImage.cs
@@ -48,7 +48,7 @@ internal sealed class ElfExecutableImage : IExecutableImage
         }
     }
 
-    private PatchResult<long> ToFileOffset(uint virtualAddress, int length)
+    private PatchResult<long> ToFileOffset(ulong virtualAddress, int length)
     {
         foreach (var seg in _loads)
         {
@@ -66,7 +66,7 @@ internal sealed class ElfExecutableImage : IExecutableImage
         return PatchResult<long>.Fail($"Virtual address 0x{virtualAddress:X8} is not mapped by any PT_LOAD segment.");
     }
 
-    public PatchResult<byte[]> ReadAtVirtualAddress(uint virtualAddress, int length)
+    public PatchResult<byte[]> ReadAtVirtualAddress(ulong virtualAddress, int length)
     {
         if (length <= 0)
             return PatchResult<byte[]>.Fail($"Invalid length: {length}");
@@ -90,7 +90,7 @@ internal sealed class ElfExecutableImage : IExecutableImage
         }
     }
 
-    public PatchResult WriteAtVirtualAddress(uint virtualAddress, byte[] bytes)
+    public PatchResult WriteAtVirtualAddress(ulong virtualAddress, byte[] bytes)
     {
         if (bytes == null || bytes.Length == 0)
             return PatchResult.Fail("Bytes cannot be null or empty");
